gdog/requirements.txt
aredis
asyncpg
aio-pika
sanic
sanic-jwt
aiohttp
Sanic-Cors
requests
ujson
pillow
locustio